refactored Chatwindow into panels
